EDGEAI-773: Refactor benchmarks to use shared common module
- Extract shared utilities to crates/image/benches/common.rs:
  - Hardware detection (dma_available, g2d_available, opengl_available)
  - Test data loading (find_testdata_path, get_test_data)
  - BenchConfig struct with id() and throughput() methods
  - calculate_letterbox() and format_name() helpers
- Update image_benchmark.rs and pipeline_benchmark.rs to use common module
- Add sonar.coverage.exclusions for benchmark files
- Reduce code duplication by ~400 lines
