+++ sonar-project.properties
@@ -35,6 +35,9 @@ sonar.objc.file.suffixes=-
 # Exclusions (build artifacts, external code, generated files)
 sonar.exclusions=**/target/**,**/build/**,**/venv/**,**/__pycache__/**,**/testdata/**,**/*.lock
 
+# Coverage exclusions (benchmark code doesn't need test coverage)
+sonar.coverage.exclusions=**/benches/**/*.rs
+
 # Quality gate settings
 # Note: Quality gate is checked via PR decoration, not workflow failure.
 # This prevents blocking merges while still showing quality gate status.
